Break ResmiDao into two levels of abstraction
