+++ resmi/src/test/java/io/corbel/resources/rem/service/DefaultResmiServiceTest.java
@@ -5,9 +5,9 @@ import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonPrimitive;
 import io.corbel.lib.queries.request.*;
+import io.corbel.resources.rem.dao.MongoResmiDao;
 import io.corbel.resources.rem.dao.NotFoundException;
 import io.corbel.resources.rem.dao.RelationMoveOperation;
-import io.corbel.resources.rem.dao.ResmiDao;
 import io.corbel.resources.rem.model.ResourceUri;
 import io.corbel.resources.rem.request.CollectionParameters;
 import io.corbel.resources.rem.request.RelationParameters;
@@ -50,7 +50,8 @@ import static org.mockito.Mockito.*;
 
     String RELATION_URI = "RELATION_URI";
 
-    @Mock ResmiDao resmiDao;
+    @Mock
+    MongoResmiDao resmiDao;
     CollectionParameters collectionParametersMock;
     @Mock ResourceQuery resourceQueryMock;
     @Mock List<ResourceQuery> resourceQueriesMock;
